configure: drop support for Jade/OpenJade/Saxon XSLT processors
Drop configure support for Jade/OpenJade XSLT processors.
Support is not functional nor is it clear it ever worked at all.

git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@31535 b3059339-0415-0410-9bf9-f77b7e298cf2

Drop support for the Saxon XSLT processor.
The XSLT processor detection needs to be overhauled and simplified. Since
nobody appears to be using Saxon it is just a burden that hinders the
simplification. It can be added back cleanly if there is popular request.

git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@31536 b3059339-0415-0410-9bf9-f77b7e298cf2