vd_qtvideo: fix crash on OS X with QuickTime
Do not call Setup_FS_Segment if the QuickTime framework is used on OSX
for decoding, even if the loader code is compiled in.
There is no point in it and since Setup_LDT_Keeper is not called before
it will actually crash on OSX due to the auto-alloc functionality not
being initialized and thus it will try to set fs to 0xffffffff.

git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@32486 b3059339-0415-0410-9bf9-f77b7e298cf2