command.c: use different field in property table for integer data
The video equalizer properties plus tv_* and teletext_* ones use a
single function to handle multiple properties, with data in the struct
m_option "prop" argument indicating which property is being
handled. They all use integer data for that, but stored it in the
"priv" field of the struct which has type "void *". This caused
warnings and is not standard C (it's not guaranteed that you can cast
an arbitrary integer to "void *' and back - it'd work in the _other_
direction with intptr_t if that type exists).

Change the code to store the data in the "offset" field of the prop
struct instead. The name is not optimal for the way the functions use
it (except for gamma), but it'll do. (An alternative would be to store
a pointer in the priv field as in "(void *)&(const int){123}", but
that's somewhat ugly - the explicit (void *) cast is needed to avoid a
qualifier warning unless make it non-const data).