build: cygwin: don't rely on _WIN32 being defined
configure: Compilation fixes for current Cygwin

git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@32724 b3059339-0415-0410-9bf9-f77b7e298cf2

Do not #define _WIN32 on the command line for Cygwin.

Newer Cygwin versions no longer do this and hopefully we should be able
to survive without this hack as well.  This change necessitates adapting
two #ifdefs in the MPlayer codebase.  It is committed untested as I do
not have access to a Cygwin system.

git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@32763 b3059339-0415-0410-9bf9-f77b7e298cf2