options: mark -fs with CONF_NOSAVE to keep state between files
Most of the time users want to keep current fullscreen state when
switching to another file, rather than always resetting to the global
default that was determined when the command line was parsed. Add the
CONF_NOSAVE flag to the -fs option to achieve this. The change means
that the fullscreen state from the previous file is kept unless there
is an explicit file-specific -fs or -nofs setting for the new file
(or in case file groups are used, moving to the new file enters a
group with explicitly specified state).