configure: use LINGUAS environment variable for messages
Previously the default with --enable-translation was to install all
language files if --language-msg was not explicitly specified. Change
the default to the languages set with the --language option or LINGUAS
environment variable if those are set; if neither is set then the
default is still to install all message languages.