stream_cdda: work around libcdparanoia caching issues
Constrain libcdparanoia's caching which badly breaks playback with
-nocache (libcdparanoia requests a huge chunk at once, then lets the
disk spin down while that is being played, leading to a pause when the
disk needs to spin back up after the big chunk is finished). Switching
to libcdio by default which does not have this ssue might be a better
long-term solution though.

git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@33557 b3059339-0415-0410-9bf9-f77b7e298cf2